services/background-worker/requirements.txt
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
minio==7.2.0
python-dotenv==1.0.0
prometheus-client==0.19.0
